Streamlit and local command
To save the transcript locally we need to run the command 'python main.py'. It will fetch the episode id from the ListenNotesAPI and use the AssemblyAI API for the transcribing and polling. To see the following project in real life we need to make a web application for the same. Will use streamlit to make the application. The application can be run using the command 'streamlit run main.py'